requirements.txt
Flask==0.8
Flask-OAuth==0.11
oauth2==1.5.211
Flask-SQLAlchemy==0.15